Ensure systemctl daemon-reload is before services
This collector ensures that if a systemctl daemon-reload is triggered
that it happens *before* we refresh the service. Otherwise the service
may refresh but not pick up changes from the unit file and then never
refresh again. Because we don't manage the service resource we use a
collector to impose this ordering.

Fixes #74