Ensure systemctl daemon-reload is before services
This collector ensures that if a systemctl daemon-reload is triggered
that it happens *before* we refresh the service. Otherwise the service
may refresh but not pick up changes from the unit file and then never
refresh again. Because we don't manage the service resource we use a
collector to impose this ordering.

Fixes #74

Only collecting on services that should be running to prevent a
dependency loop in unit_file if we are removing the service. I think
this is okay since we only really care about the timing of the
daemon-reload if we are about to start or restart the service in which
case ensure must be running/true.